Fix for test 7.9.6.a.

+++ src/java/org/orbeon/oxf/xforms/function/SecondsFromDateTime.java
@@ -17,7 +17,7 @@ import org.orbeon.saxon.expr.XPathContext;
 import org.orbeon.saxon.om.Item;
 import org.orbeon.saxon.trans.XPathException;
 import org.orbeon.saxon.value.DateTimeValue;
-import org.orbeon.saxon.value.Int64Value;
+import org.orbeon.saxon.value.DecimalValue;
 import org.orbeon.saxon.value.StringValue;
 
 import java.math.BigDecimal;
@@ -34,7 +34,7 @@ public class SecondsFromDateTime extends XFormsFunction {
 
             // "the return value is equal to the number of seconds difference between the specified dateTime
             // (normalized to UTC) and 1970-01-01T00:00:00Z"
-            return new Int64Value(value.normalize(context).toJulianInstant().subtract(BASELINE).longValue());
+            return new DecimalValue(value.normalize(context).toJulianInstant().subtract(BASELINE));
 
         } catch (XPathException e) {
             return NAN;
